Rename restoretarget to relativetime, use relative time in snapshot create delete at
